requirements.txt
beautifulsoup4==4.7.1
pytest==4.4.1
requests-toolbelt==0.9.1
Flask==1.0.2
google-cloud-datastore